+++ junit/src/main/java/io/cucumber/junit/JUnitCucumberOptionsProvider.java
@@ -27,8 +27,7 @@ final class JUnitCucumberOptionsProvider implements CucumberOptionsAnnotationPar
             String name = clazzAnnotation.annotationType().getName();
             if ("io.cucumber.testng.CucumberOptions".equals(name)) {
                 log.warn(() -> "Ignoring options provided by " + name + " on " + clazz.getName() + ". " +
-                        "It is recommend to use separate runner classes for JUnit and TestNG."
-                );
+                        "It is recommend to use separate runner classes for JUnit and TestNG.");
             }
         }
     }